feat: detect account API block after 3 consecutive no-AJAX cycles and cooldown 3h
- Track _consecutiveNoAjaxCycles counter in unCheckCycle()
- If all facilities return no AJAX response for 3 cycles in a row,
  return ACCOUNT_BLOCKED signal from the cycle
- Handle ACCOUNT_BLOCKED in _runLoop() with prominent error logging,
  3-hour sleep, counter reset, and re-login before resuming
- Reset counter immediately when any facility returns a real AJAX response
- Improve no-AJAX fallback warning log (was silent debug, now warn)

+++ database.js
@@ -31,7 +31,7 @@ function initDatabase() {
       facility_ids TEXT NOT NULL DEFAULT '[]',
       start_date TEXT NOT NULL,
       end_date TEXT NOT NULL,
-      check_interval_seconds INTEGER NOT NULL DEFAULT 30,
+      check_interval_seconds REAL NOT NULL DEFAULT 30,
       auto_book INTEGER NOT NULL DEFAULT 1,
       max_relogin_attempts INTEGER NOT NULL DEFAULT 5,
       request_timeout_ms INTEGER NOT NULL DEFAULT 20000,
